// src/components/SearchForm/api/airports/SearchCity.ts
export interface SearchCity {
  cityCode: string;
  cityName: string;
  countryCode: string;
  countryName: string;
  createdAt: string;
  displayLocation: string;
  displayName: string;
  geopoint: string;
  iataCode: string;
  id: string;
  latitude: string;
  longitude: string;
  name: string;
  region: string;
  regionCode: string;
  searchText: string;
  shortRegionCode: string;
  timezone: string;
  timeDifference: string;
  updatedAt: string;
  type?: string;
}